:rocket: Selection mapping, payload validation
This commit includes utility functions for mapping a Selection, and numerous improvements to internal data handling.

I've streamlined GraphQLResult structure so that it more accurately represents the return value.

Field accessors may now fail to reflect the possibility of a bad payload. This is a huge improvement over the previous approach since we no longer trigger a fatal error whenever the data is corrupt, but instead throw a bad payload error.

Additionally, we require that query selection is optional since it might fail. To prevent endless nullability wrappings, I've created a utility send function that wrapps selection in `nonNullOrFail` selection.