refactor: Combine.first -> RxSwift.take(1).asSingle().asObservable()
This refactor is a little more unusual.
Combine seems to have some way of guaranteeing that `first()` won't throw an error.
It's not clear to me how this is possible since an event stream can't know if there will be one or many events published.
That decision is made by the publisher.
The default behavior(s) for RxSwift are:
    - return an optional (for when no value is published before the stream closes)
    - throw an error (for when too many or too few events are published)

Beyond this, RxSwift has some non-Linux supported functionalities around their Single type that allows for an `async throws` retrieval of the value. We can't use that functionality here though.